Support imports of stdlib and other Grain files, as well as compiling the Grain files in-place
This adds a parser to the loader that can pull out and bundle up
both imports from the stdlib and from your own multi-file Grain projects.
It's fairly non-performant at the moment but allows you to load bigger programs.
It tries to load the dependencies in the same way that the runtime's locators do.

The loader has also changed to compiling the Grain files in-place like the Grain compiler does rather
than working based off the source being loaded.